pyproject.toml
[tool.poetry]
name = "st-app-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"
streamlit = "^1.11.0"
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.961"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
